FEATURE Fetch & Store Pokemon
DESCRIPTION
Create the rake task which imports 150 Pokemons from the official Pokemon API into our DB.

NOTES
The Rake task has an argument that can be used to customize the number of Pokemons imported.
A comment with how to run the rake task can be found into the imports.rake file.

ADDITIONAL NOTE
In this commit I've also developed all the rspecs related to the models created in the previous commit, because I've used them in the Rake task. I've also improved some models.